// src/Utilites/Types/Types.ts
export type RootStackParamList = {
  Home: undefined;
  Login: undefined;
  Location: undefined;
  Task: undefined;
  navigate: Function;
  Training: undefined;
  AlterScreen: undefined;
  OperatorHome: undefined;
  LocationManagment: undefined;
  QualityManagement: undefined;
  SurveyManagment: undefined;
  TrainingManagment: undefined;
  CleanerWorkDetails: undefined;
  Cleaners: undefined;
  CleanerList: undefined;
  ReportManagement: undefined;
};